Refactor/restructure code for groups and scene tree
use a more high-level helper file for everything related to groups
